--- a/benchmark/partial_tweets/iter.h
+++ b/benchmark/partial_tweets/iter.h
@@ -47,35 +47,35 @@ simdjson_really_inline bool Iter::Run(const padded_string &json) {
     tweet tweet;
 
     if (!iter.start_object()   || !iter.find_field_raw("created_at")) { return false; }
-    tweet.created_at = iter.get_raw_json_string().value().unescape(iter);
+    tweet.created_at = iter.consume_raw_json_string().value().unescape(iter);
 
     if (!iter.has_next_field() || !iter.find_field_raw("id")) { return false; }
-    tweet.id = iter.get_uint64();
+    tweet.id = iter.consume_uint64();
 
     if (!iter.has_next_field() || !iter.find_field_raw("text")) { return false; }
-    tweet.text = iter.get_raw_json_string().value().unescape(iter);
+    tweet.text = iter.consume_raw_json_string().value().unescape(iter);
 
     if (!iter.has_next_field() || !iter.find_field_raw("in_reply_to_status_id")) { return false; }
     if (!iter.is_null()) {
-      tweet.in_reply_to_status_id = iter.get_uint64();
+      tweet.in_reply_to_status_id = iter.consume_uint64();
     }
 
     if (!iter.has_next_field() || !iter.find_field_raw("user")) { return false; }
     {
       if (!iter.start_object()   || !iter.find_field_raw("id")) { return false; }
-      tweet.user.id = iter.get_uint64();
+      tweet.user.id = iter.consume_uint64();
 
       if (!iter.has_next_field() || !iter.find_field_raw("screen_name")) { return false; }
-      tweet.user.screen_name = iter.get_raw_json_string().value().unescape(iter);
+      tweet.user.screen_name = iter.consume_raw_json_string().value().unescape(iter);
 
       if (iter.skip_container()) { return false; } // Skip the rest of the user object
     }
 
     if (!iter.has_next_field() || !iter.find_field_raw("retweet_count")) { return false; }
-    tweet.retweet_count = iter.get_uint64();
+    tweet.retweet_count = iter.consume_uint64();
 
     if (!iter.has_next_field() || !iter.find_field_raw("favorite_count")) { return false; }
-    tweet.favorite_count = iter.get_uint64();
+    tweet.favorite_count = iter.consume_uint64();
 
     tweets.push_back(tweet);