This allows you to try to parse a value multiple ways (for example, if `value.get_string()` fails, you can still try `value.get_double()`).

It also uses standard iterator code for parsing instead of special code in value, and fixes issues where we were advancing structural indexes multiple times when you parsed values at the root. (Or attempts to ... tests are still pending.)

## Performance

By my measure, this improves things on gcc by a small amount (2% or so for PartialTweets and Kostya, both of which have skipping), and reduces instruction count slightly as well. On clang, however, it makes things a bit *worse* by about 5%, significantly increasing branch misses. I'm not sure why yet. I'll think on this one.

### Haswell gcc10.2 (Skylake)

With this patch:

```
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
PartialTweets<OnDemand>      175706 ns       175707 ns         3979           1.561k           3.79042G               0        57.975k     615.24k             0.974229          6.00211k       3.69274G          2.11194M                    3.34425                     3.43271           600.211k    1.68753k   631.515k        3.3473G/s   0.017341   57.8757k   619.048k        0.980258    5.6913k/s 3.52319G/s     2.11194M               3.34425                 3.4116        100        569.13k/s [best: throughput=  3.79 GB/s doc_throughput=  6002 docs/s instructions=     2111941 cycles=      615240 branch_miss=    1561 cache_miss=       0 cache_ref=     57975 items=       100 avg_time=    167739 ns]
PartialTweets<Iter>          413260 ns       413189 ns         1690           3.018k           1.56446G               0        59.203k    1.49034M              2.35995          2.47732k       3.69205G          4.44231M                    7.03437                     2.98073           247.732k     3.1855k   631.515k       1.42343G/s  0.0171598   59.3027k   1.49523M         2.36769    2.4202k/s 3.61876G/s     4.44231M               7.03437                2.97098        100        242.02k/s [best: throughput=  1.56 GB/s doc_throughput=  2477 docs/s instructions=     4442308 cycles=     1490342 branch_miss=    3018 cache_miss=       0 cache_ref=     59203 items=       100 avg_time=    405266 ns]
LargeRandom<OnDemand>      59208577 ns     59196413 ns           12         867.944k           777.982M        5.62205M       8.02476M     218.12M              4.74186           16.9131       3.68908G          616.695M                    13.4068                     2.82732           16.9131M    867.787k   45.9988M       741.056M/s   5.66668M   8.02553M   218.321M         4.74624    16.8929/s 3.68808G/s     616.695M               13.4068                2.82471      1000k       16.8929M/s [best: throughput=  0.78 GB/s doc_throughput=    16 docs/s instructions=   616695084 cycles=   218119787 branch_miss=  867944 cache_miss= 5622048 cache_ref=   8024758 items=   1000000 avg_time=  59196416 ns]
LargeRandom<Iter>          53297188 ns     53287207 ns           13         870.255k            864.43M          5.628M       8.01742M    196.306M              4.26764           18.7924       3.68908G          570.695M                    12.4067                     2.90716           18.7924M    870.999k   45.9988M       823.234M/s   5.65342M   8.01889M    196.51M         4.27207    18.7662/s 3.68775G/s     570.695M               12.4067                2.90415      1000k       18.7662M/s [best: throughput=  0.86 GB/s doc_throughput=    18 docs/s instructions=   570694598 cycles=   196306411 branch_miss=  870255 cache_miss= 5628004 cache_ref=   8017425 items=   1000000 avg_time=  53285155 ns]
Kostya<OnDemand>           59404278 ns     59392435 ns           12         451.256k           2.31519G        9.95922M       13.9063M    218.733M              1.59304           16.8617        3.6882G          648.831M                    4.72547                     2.96632           8.84038M    451.521k   137.305M       2.15306G/s   10.0041M   13.9063M   218.971M         1.59477    16.8372/s 3.68684G/s     648.831M               4.72547                 2.9631   524.288k       8.82752M/s [best: throughput=  2.32 GB/s doc_throughput=    16 docs/s instructions=   648830818 cycles=   218732563 branch_miss=  451256 cache_miss= 9959225 cache_ref=  13906324 items=    524288 avg_time=  59392557 ns]
Kostya<Iter>               57719889 ns     57719511 ns           12         452.567k            2.3822G        9.99273M       13.8838M    212.574M              1.54818           17.3497       3.68808G          640.967M                     4.6682                     3.01527           9.09623M    452.514k   137.305M       2.21546G/s   10.0028M   13.8939M    212.85M          1.5502    17.3252/s 3.68766G/s     640.967M                4.6682                3.01136   524.288k       9.08338M/s [best: throughput=  2.38 GB/s doc_throughput=    17 docs/s instructions=   640966891 cycles=   212573506 branch_miss=  452567 cache_miss= 9992732 cache_ref=  13883818 items=    524288 avg_time=  57708281 ns]
```

Before this patch:

```
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
PartialTweets<OnDemand>      180978 ns       180977 ns         3869           1.564k           3.70601G               0        58.491k    629.234k             0.996388          5.86844k       3.69262G          2.11195M                    3.34425                     3.35638           586.844k     1.7594k   631.515k       3.24983G/s   1.29232m   58.6616k   638.611k         1.01124   5.52557k/s 3.52869G/s     2.11195M               3.34425                 3.3071        100       552.557k/s [best: throughput=  3.71 GB/s doc_throughput=  5868 docs/s instructions=     2111947 cycles=      629234 branch_miss=    1564 cache_miss=       0 cache_ref=     58491 items=       100 avg_time=    173042 ns]
PartialTweets<Iter>          386232 ns       386152 ns         1815           3.108k           1.67989G               0        59.344k    1.38794M               2.1978           2.6601k       3.69206G          4.40453M                    6.97454                     3.17342            266.01k    3.27626k   631.515k       1.52309G/s   0.015427   59.3538k   1.39539M          2.2096   2.58965k/s 3.61358G/s     4.40453M               6.97454                3.15648        100       258.965k/s [best: throughput=  1.68 GB/s doc_throughput=  2660 docs/s instructions=     4404526 cycles=     1387942 branch_miss=    3108 cache_miss=       0 cache_ref=     59344 items=       100 avg_time=    378292 ns]
LargeRandom<OnDemand>      58630783 ns     58618830 ns           12         869.638k           785.824M        5.62459M       8.02289M    215.945M              4.69459           17.0836       3.68912G          614.695M                    13.3633                     2.84653           17.0836M    870.219k   45.9988M       748.358M/s   5.66826M    8.0247M   216.143M         4.69889    17.0594/s 3.68727G/s     614.695M               13.3633                2.84392      1000k       17.0594M/s [best: throughput=  0.79 GB/s doc_throughput=    17 docs/s instructions=   614695180 cycles=   215945356 branch_miss=  869638 cache_miss= 5624589 cache_ref=   8022886 items=   1000000 avg_time=  58618766 ns]
LargeRandom<Iter>          54140042 ns     54139736 ns           13         864.127k           850.741M        5.60486M       8.02276M    199.446M              4.33591           18.4949       3.68874G          571.695M                    12.4285                     2.86641           18.4949M    864.509k   45.9988M       810.271M/s   5.65229M   8.02376M   199.652M         4.34037    18.4707/s 3.68772G/s     571.695M               12.4285                2.86346      1000k       18.4707M/s [best: throughput=  0.85 GB/s doc_throughput=    18 docs/s instructions=   571694596 cycles=   199446414 branch_miss=  864127 cache_miss= 5604864 cache_ref=   8022761 items=   1000000 avg_time=  54128220 ns]
Kostya<OnDemand>           60838777 ns     60839545 ns           11         453.749k           2.26086G        10.0671M        14.032M    224.018M              1.63154            16.466       3.68867G          647.258M                    4.71402                     2.88931            8.6329M    454.131k   137.305M       2.10184G/s   10.0455M   14.0312M   224.343M          1.6339    16.4367/s 3.68746G/s     647.258M               4.71402                2.88513   524.288k       8.61755M/s [best: throughput=  2.26 GB/s doc_throughput=    16 docs/s instructions=   647257988 cycles=   224017971 branch_miss=  453749 cache_miss=10067083 cache_ref=  14031971 items=    524288 avg_time=  60826928 ns]
Kostya<Iter>               59252275 ns     59250146 ns           12          453.17k           2.32348G        10.0747M        14.024M    217.951M              1.58735            16.922       3.68816G          639.918M                    4.66056                     2.93607             8.872M    452.788k   137.305M       2.15823G/s   10.0237M   14.0261M   218.466M          1.5911    16.8776/s 3.68719G/s     639.918M               4.66056                2.92914   524.288k       8.84872M/s [best: throughput=  2.32 GB/s doc_throughput=    16 docs/s instructions=   639918058 cycles=   217950758 branch_miss=  453170 cache_miss=10074750 cache_ref=  14023984 items=    524288 avg_time=  59240277 ns]
```

## Haswell clang10 (Skylake)

With this patch:

```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
PartialTweets<OnDemand>     173875 ns       173876 ns         4093           1.404k           3.83827G               0        54.655k    607.533k             0.962025          6.07788k       3.69251G          1.93361M                    3.06186                     3.18272           607.788k       1.58k   631.515k       3.38255G/s  0.0163694   54.8714k   612.776k        0.970328   5.75123k/s 3.52422G/s     1.93361M               3.06186                3.15549        100       575.123k/s [best: throughput=  3.84 GB/s doc_throughput=  6077 docs/s instructions=     1933608 cycles=      607533 branch_miss=    1404 cache_miss=       0 cache_ref=     54655 items=       100 avg_time=    166032 ns]
PartialTweets<Iter>         179035 ns       179038 ns         3916           2.823k           3.75803G               0        54.568k    620.502k             0.982561          5.95082k        3.6925G          1.83669M                    2.90839                     2.96001           595.082k     3.1604k   631.515k       3.28503G/s  0.0735444   54.7074k   631.576k          1.0001   5.58542k/s 3.52761G/s     1.83669M               2.90839                2.90811        100       558.542k/s [best: throughput=  3.76 GB/s doc_throughput=  5950 docs/s instructions=     1836691 cycles=      620502 branch_miss=    2823 cache_miss=       0 cache_ref=     54568 items=       100 avg_time=    171125 ns]
Creating a source file spanning 44921 KB 
LargeRandom<OnDemand>     68934797 ns     68935741 ns           10         920.172k           667.883M         5.6818M       7.91918M    254.089M              5.52382           14.5196       3.68927G          678.851M                     14.758                      2.6717           14.5196M    920.914k   45.9988M       636.359M/s   5.72998M   7.92078M   254.267M          5.5277    14.5063/s 3.68847G/s     678.851M                14.758                2.66983      1000k       14.5063M/s [best: throughput=  0.67 GB/s doc_throughput=    14 docs/s instructions=   678850849 cycles=   254089075 branch_miss=  920172 cache_miss= 5681804 cache_ref=   7919178 items=   1000000 avg_time=  68922499 ns]
LargeRandom<Iter>         60783233 ns     60784064 ns           12         962.729k           758.008M         5.6942M       7.90349M    223.878M              4.86704           16.4789       3.68926G           581.85M                    12.6493                     2.59896           16.4789M    968.679k   45.9988M         721.7M/s   5.72905M   7.90575M   224.184M          4.8737    16.4517/s 3.68821G/s      581.85M               12.6493                2.59541      1000k       16.4517M/s [best: throughput=  0.76 GB/s doc_throughput=    16 docs/s instructions=   581850353 cycles=   223878024 branch_miss=  962729 cache_miss= 5694202 cache_ref=   7903490 items=   1000000 avg_time=  60770921 ns]
Creating a source file spanning 134087 KB 
Kostya<OnDemand>          61465597 ns     61466425 ns           11         455.813k           2.23602G        10.1644M       14.0161M    226.542M              1.64992           16.2851       3.68925G          649.466M                    4.73009                     2.86687           8.53806M     456.19k   137.305M       2.08041G/s   10.1823M   14.0158M   226.699M         1.65106     16.269/s 3.68817G/s     649.466M               4.73009                2.86488   524.288k       8.52966M/s [best: throughput=  2.24 GB/s doc_throughput=    16 docs/s instructions=   649465616 cycles=   226542030 branch_miss=  455813 cache_miss=10164365 cache_ref=  14016146 items=    524288 avg_time=  61452878 ns]
Kostya<Iter>              58592324 ns     58593125 ns           12          463.54k           2.34617G        10.1423M       13.9871M    215.897M              1.57239           17.0873       3.68909G            604.9M                    4.40552                      2.8018           8.95866M    465.007k   137.305M       2.18243G/s   10.1704M   13.9877M   216.112M         1.57396    17.0668/s 3.68835G/s       604.9M               4.40552                2.79901   524.288k       8.94794M/s [best: throughput=  2.35 GB/s doc_throughput=    17 docs/s instructions=   604900426 cycles=   215897094 branch_miss=  463540 cache_miss=10142262 cache_ref=  13987106 items=    524288 avg_time=  58579940 ns]
```

Before this patch:

```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
…