@jkeiser 2020-09-21
Before making any decisions, the two avenues to explore here are:

- [ ] Try different versions of get_string/etc. for `&` vs. `&&`.

  Before this patch, we only supported `&&`, which meant the value was scheduled to be destroyed anyway and we could release the iterator before even trying to parse the JSON. But supporting this means supporting `&` as well, so we now release the iterator only if parsing succeeds. If clang isn't doing a good job optimizing that together with the destructor (which releases the iterator if it wasn't already released), it would explain these results pretty clearly.
- [ ] Check whether the changes to `document` (particularly the addition of a non-default move constructor) are a factor