@jkeiser 2020-09-22
I tried a quick related experiment where I moved `const uint8_t *json` from value/document over into the json_iterator (under the theory that if the compiler was failing to identify when the value was live, it might keep more than one around at a time). It slowed things down appreciably, especially on clang.

Neither here nor here, but I figured I'd write down the experiment, anyway :)