@jkeiser 2020-09-26
OK, having an && variant fixed the performance drop. Merging.